fix bug not build vercel
